requirements.txt
streamlit
openai==1.35.15
python-dotenv
httpx==0.27.2
